// web/src/api/models/bot.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BotConfig } from './bot-config';

/**
 * 
 * @export
 * @interface Bot
 */
export interface Bot {
    /**
     * 
     * @type {string}
     * @memberof Bot
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Bot
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Bot
     */
    'description'?: string;
    /**
     * The type of bot
     * @type {string}
     * @memberof Bot
     */
    'type'?: BotTypeEnum;
    /**
     * 
     * @type {BotConfig}
     * @memberof Bot
     */
    'config'?: BotConfig;
    /**
     * 
     * @type {string}
     * @memberof Bot
     */
    'created'?: string;
    /**
     * 
     * @type {string}
     * @memberof Bot
     */
    'updated'?: string;
}

export const BotTypeEnum = {
    knowledge: 'knowledge',
    common: 'common',
    agent: 'agent'
} as const;

export type BotTypeEnum = typeof BotTypeEnum[keyof typeof BotTypeEnum];