// web/src/api/models/chat.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ATRAG API
 * ATRAG API Documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Chat
 */
export interface Chat {
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'title'?: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'bot_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'peer_id'?: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'peer_type'?: ChatPeerTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'status'?: ChatStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'created'?: string;
    /**
     * 
     * @type {string}
     * @memberof Chat
     */
    'updated'?: string;
}

export const ChatPeerTypeEnum = {
    system: 'system',
    feishu: 'feishu',
    weixin: 'weixin',
    weixin_official: 'weixin_official',
    web: 'web',
    dingtalk: 'dingtalk'
} as const;

export type ChatPeerTypeEnum = typeof ChatPeerTypeEnum[keyof typeof ChatPeerTypeEnum];
export const ChatStatusEnum = {
    active: 'active',
    archived: 'archived'
} as const;

export type ChatStatusEnum = typeof ChatStatusEnum[keyof typeof ChatStatusEnum];